+++ actix-http/src/encoding/encoder.rs
@@ -131,18 +131,9 @@ where
                     Poll::Ready(Some(Ok(std::mem::take(b))))
                 }
             }
-            // TODO: MSRV 1.51: poll_map_err
-            EncoderBodyProj::Stream(b) => match ready!(b.poll_next(cx)) {
-                Some(Err(err)) => Poll::Ready(Some(Err(EncoderError::Body(err)))),
-                Some(Ok(val)) => Poll::Ready(Some(Ok(val))),
-                None => Poll::Ready(None),
-            },
+            EncoderBodyProj::Stream(b) => b.poll_next(cx).map_err(EncoderError::Body),
             EncoderBodyProj::BoxedStream(ref mut b) => {
-                match ready!(b.as_pin_mut().poll_next(cx)) {
-                    Some(Err(err)) => Poll::Ready(Some(Err(EncoderError::Boxed(err)))),
-                    Some(Ok(val)) => Poll::Ready(Some(Ok(val))),
-                    None => Poll::Ready(None),
-                }
+                b.as_pin_mut().poll_next(cx).map_err(EncoderError::Boxed)
             }
         }
     }